requirements.txt
torch
transformers
sentence-transformers
langchain
chromadb
streamlit
Pillow
peft
langchain-community